+++ mock/src/crypto.rs
@@ -1,6 +1,6 @@
 use aleph_bft_types::{
     Index, KeyBox as KeyBoxT, MultiKeychain as MultiKeychainT, NodeCount, NodeIndex,
-    PartialMultisignature as PartialMultisignatureT, SignatureSet,
+    PartialMultisignature as PartialMultisignatureT,
 };
 use async_trait::async_trait;
 use codec::{Decode, Encode};
@@ -70,57 +70,5 @@ impl MultiKeychainT for KeyBox {
     }
     fn is_complete(&self, _: &[u8], partial: &Self::PartialMultisignature) -> bool {
         (self.count * 2) / 3 < NodeCount(partial.signed_by.len())
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Encode, Decode)]
-pub struct VerboseSignature {
-    msg: Vec<u8>,
-    index: NodeIndex,
-}
-
-impl VerboseSignature {
-    pub fn new(msg: Vec<u8>, index: NodeIndex) -> Self {
-        Self { msg, index }
-    }
-}
-
-pub type VerbosePartialMultisignature = SignatureSet<VerboseSignature>;
-
-#[derive(Clone, Debug)]
-pub struct VerboseKeyBox {
-    count: NodeCount,
-    index: NodeIndex,
-}
-
-impl VerboseKeyBox {
-    pub fn new(count: NodeCount, index: NodeIndex) -> Self {
-        VerboseKeyBox { count, index }
-    }
-}
-
-impl Index for VerboseKeyBox {
-    fn index(&self) -> NodeIndex {
-        self.index
-    }
-}
-
-#[async_trait]
-impl KeyBoxT for VerboseKeyBox {
-    type Signature = VerboseSignature;
-
-    fn node_count(&self) -> NodeCount {
-        self.count
-    }
-
-    async fn sign(&self, msg: &[u8]) -> Self::Signature {
-        VerboseSignature {
-            msg: msg.to_vec(),
-            index: self.index,
-        }
-    }
-
-    fn verify(&self, msg: &[u8], sgn: &Self::Signature, index: NodeIndex) -> bool {
-        index == sgn.index && msg == sgn.msg
     }
 }
