+++ mock/src/lib.rs
@@ -1,5 +1,5 @@
-mod constant_signature_keybox;
 mod bad_verbose_multikeychain;
+mod constant_signature_keybox;
 mod crypto;
 mod dataio;
 mod hasher;
@@ -8,13 +8,12 @@ mod signable_byte;
 mod spawner;
 mod threshold_multi_keychain;
 mod unreliable_router;
+mod verbose_keybox;
+mod verbose_signature;
 
-pub use constant_signature_keybox::ConstantSignatureKeyBox;
 pub use bad_verbose_multikeychain::BadVerboseMultiKeychain;
-pub use crypto::{
-    KeyBox, PartialMultisignature, Signature, VerboseKeyBox, VerbosePartialMultisignature,
-    VerboseSignature,
-};
+pub use constant_signature_keybox::ConstantSignatureKeyBox;
+pub use crypto::{KeyBox, PartialMultisignature, Signature};
 pub use dataio::{Data, DataProvider, FinalizationHandler};
 pub use hasher::{Hash64, Hasher64};
 pub use network::{Network, NetworkReceiver, NetworkSender};
@@ -22,3 +21,5 @@ pub use signable_byte::SignableByte;
 pub use spawner::Spawner;
 pub use threshold_multi_keychain::ThresholdMultiKeychain;
 pub use unreliable_router::{configure_network, NetworkHook, Peer, UnreliableRouter};
+pub use verbose_keybox::VerboseKeyBox;
+pub use verbose_signature::{VerbosePartialMultisignature, VerboseSignature};
